backend/requirements.txt
flask
flask-cors
ibm-watson
ibm-cloud-sdk-core
python-dotenv
cloudant
